docs: CI builds full app + artifact; fix WriteInt32BE cast, F1 key

+++ protocol/framing.h
@@ -39,8 +39,8 @@ inline uint32_t ReadInt32BE(const uint8_t* p) {
 // Writes a 4-byte big-endian unsigned integer.
 inline void WriteInt32BE(uint8_t* p, uint32_t v) {
     p[0] = uint8_t(v >> 24);
-    p[1] = uint32_t(v >> 16);
-    p[2] = uint32_t(v >> 8);
+    p[1] = uint8_t(v >> 16);
+    p[2] = uint8_t(v >> 8);
     p[3] = uint8_t(v);
 }
